Prefetch documentation for next / previous completion in selector
In #21286, documentation fetch was made more efficient by only
fetching the current completion. However, this causes some delay for
documentation to appear when navigating the completions. This change
somewhat mitigates this by prefetching the documentation for the
next / previous completions in the list.

It also makes this fetch a bit more efficient by attempting to skip
fetch for completions that are known to have documentation.